+++ bfdd/dplane.c
@@ -354,7 +354,7 @@ bfd_dplane_session_state_change(struct bfd_dplane_ctx *bdc,
 	bs->remote_timers.required_min_echo = ntohl(state->required_echo_rx);
 
 	/* Notify and update counters. */
-	control_notify(bs, bs->ses_state);
+	ptm_bfd_notify(bs, bs->ses_state);
 
 	/* No state change. */
 	if (old_state == bs->ses_state)
