+++ bfdd/ptm_adapter.c
@@ -148,7 +148,6 @@ static void _ptm_bfd_session_del(struct bfd_session *bs, uint8_t diag)
 				"ptm-del-session: [%s] session refcount is zero but it was configured by CLI",
 				bs_to_string(bs));
 		} else {
-			control_notify_config(BCM_NOTIFY_CONFIG_DELETE, bs);
 			bfd_session_free(bs);
 		}
 	}
@@ -892,7 +891,7 @@ static struct ptm_client *pc_new(uint32_t pid)
 		return pc;
 
 	/* Allocate the client data and save it. */
-	pc = XCALLOC(MTYPE_BFDD_CONTROL, sizeof(*pc));
+	pc = XCALLOC(MTYPE_BFDD_CLIENT, sizeof(*pc));
 
 	pc->pc_pid = pid;
 	TAILQ_INSERT_HEAD(&pcqueue, pc, pc_entry);
@@ -910,7 +909,7 @@ static void pc_free(struct ptm_client *pc)
 		pcn_free(pcn);
 	}
 
-	XFREE(MTYPE_BFDD_CONTROL, pc);
+	XFREE(MTYPE_BFDD_CLIENT, pc);
 }
 
 static void pc_free_all(void)
@@ -934,7 +933,7 @@ static struct ptm_client_notification *pcn_new(struct ptm_client *pc,
 		return pcn;
 
 	/* Save the client notification data. */
-	pcn = XCALLOC(MTYPE_BFDD_NOTIFICATION, sizeof(*pcn));
+	pcn = XCALLOC(MTYPE_BFDD_CLIENT_NOTIFICATION, sizeof(*pcn));
 
 	TAILQ_INSERT_HEAD(&pc->pc_pcnqueue, pcn, pcn_entry);
 	pcn->pcn_pc = pc;
@@ -982,5 +981,5 @@ static void pcn_free(struct ptm_client_notification *pcn)
 	pcn->pcn_pc = NULL;
 	TAILQ_REMOVE(&pc->pc_pcnqueue, pcn, pcn_entry);
 
-	XFREE(MTYPE_BFDD_NOTIFICATION, pcn);
+	XFREE(MTYPE_BFDD_CLIENT_NOTIFICATION, pcn);
 }
