+++ bfdd/subdir.am
@@ -17,8 +17,6 @@ bfdd_libbfd_a_SOURCES = \
 	bfdd/bfdd_vty.c \
 	bfdd/bfdd_cli.c \
 	bfdd/bfd_packet.c \
-	bfdd/config.c \
-	bfdd/control.c \
 	bfdd/dplane.c \
 	bfdd/event.c \
 	bfdd/ptm_adapter.c \
@@ -37,7 +35,6 @@ clippy_scan += \
 	# end
 
 noinst_HEADERS += \
-	bfdd/bfdctl.h \
 	bfdd/bfdd_nb.h \
 	bfdd/bfd.h \
 	# end
