+++ snapcraft/scripts/bfdd-service
@@ -9,6 +9,5 @@ exec $SNAP/sbin/bfdd \
     -f $SNAP_DATA/bfdd.conf \
     --pid_file $SNAP_DATA/bfdd.pid \
     --socket $SNAP_DATA/zsock \
-    --vty_socket $SNAP_DATA \
-    --bfdctl $SNAP_DATA/bfdd.sock
+    --vty_socket $SNAP_DATA
